requirements.txt
pdf2image==1.16.2
torch== 2.1.1
torchvision
torchaudio
datasets==2.10.1
paddlepaddle==2.5.2
paddleocr>=2.0.1
PyMuPDF==1.21.1

requests~=2.27.1
Flask~=3.0.2
Werkzeug~=3.0.1
numpy~=1.26.4
opencv-python~=4.9.0.80
pillow~=10.2.0
shapely~=2.0.3
six~=1.16.0
pyclipper~=1.3.0.post5
scipy~=1.12.0
scikit-image~=0.22.0
imgaug~=0.4.0
lmdb~=1.4.1
visualdl~=2.5.3
tqdm~=4.66.2
rapidfuzz~=3.6.1
Cython~=3.0.8
PyYAML~=6.0.1
setuptools~=68.2.0
paddleocr~=2.6.0.1
matplotlib~=3.8.3
openpyxl~=3.1.2
lxml~=5.1.0
premailer~=3.10.0
fitz-utils